implemented logic, adjusted tests

+++ src/main/java/com/roomsmanager/models/RoomsRevenue.java
@@ -5,11 +5,11 @@ import lombok.Data;
 @Data
 public class RoomsRevenue {
     int premiumUsage = 0;
-    float premiumRevenue = 0;
+    double premiumRevenue = 0;
     int economyUsage = 0;
-    float economyRevenue = 0;
+    double economyRevenue = 0;
 
-    public RoomsRevenue(int premiumUsage, float premiumRevenue, int economyUsage, float economyRevenue) {
+    public RoomsRevenue(int premiumUsage, double premiumRevenue, int economyUsage, double economyRevenue) {
         this.premiumUsage = premiumUsage;
         this.premiumRevenue = premiumRevenue;
         this.economyUsage = economyUsage;
